+++ csrc/engine/rdma/rdma_scheduler.h
@@ -6,7 +6,7 @@
 
 #include "engine/assignment.h"
 #include "engine/rdma/rdma_assignment.h"
-#include "engine/rdma/rdma_transport.h"
+#include "engine/rdma/rdma_context.h"
 
 namespace slime {
