+++ csrc/python/bind.cpp
@@ -1,7 +1,7 @@
 #include "engine/assignment.h"
 #include "engine/rdma/rdma_assignment.h"
 #include "engine/rdma/rdma_config.h"
-#include "engine/rdma/rdma_transport.h"
+#include "engine/rdma/rdma_context.h"
 #include <functional>
 #include <pybind11/cast.h>
 #include <pybind11/pytypes.h>
